Assignment 1/requirements.txt
Flask
python-rocksdb